requirements.txt
appdirs==1.4.4
autoflake @ file:///home/conda/feedstock_root/build_artifacts/autoflake_1665733413003/work
autopep8 @ file:///home/conda/feedstock_root/build_artifacts/autopep8_1719133654277/work
beautifulsoup4 @ file:///home/conda/feedstock_root/build_artifacts/beautifulsoup4_1738740337718/work
black==25.1.0
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1725267563793/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1739515848642/work/certifi
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1725560551097/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1728479282467/work
click==8.1.8
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
coverage==7.8.0
docformatter @ file:///home/conda/feedstock_root/build_artifacts/docformatter_1689652499088/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1733208806608/work
execnet==2.1.1
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
iniconfig @ file:///home/conda/feedstock_root/build_artifacts/iniconfig_1733223141826/work
isort==6.0.1
libsass @ file:///Users/runner/miniforge3/conda-bld/libsass_1742562609981/work
markdown-it-py @ file:///home/conda/feedstock_root/build_artifacts/markdown-it-py_1733250460757/work
mdurl @ file:///home/conda/feedstock_root/build_artifacts/mdurl_1733255585584/work
mypy-extensions==1.0.0
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1742254867999/work/dist/numpy-2.2.4-cp311-cp311-macosx_11_0_arm64.whl#sha256=10701f95f0b55ae60eb6c620ad18ef4823f33de7d2ed4c3914518ed68a58caeb
opentimestamps==0.4.5
opentimestamps-client==0.7.2
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1731802491770/work
pandas @ file:///Users/runner/miniforge3/conda-bld/pandas_1726878432242/work
pandas-stubs @ file:///home/conda/feedstock_root/build_artifacts/pandas-stubs_1728505590299/work
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
pluggy @ file:///home/conda/feedstock_root/build_artifacts/pluggy_1733222765875/work
psutil @ file:///Users/runner/miniforge3/conda-bld/psutil_1740663164053/work
pycodestyle @ file:///home/conda/feedstock_root/build_artifacts/pycodestyle_1722846760694/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pycryptodomex==3.22.0
pyflakes @ file:///home/conda/feedstock_root/build_artifacts/pyflakes_1704424584912/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
pytest @ file:///home/conda/feedstock_root/build_artifacts/pytest_1740946542080/work
pytest-cov==6.1.1
pytest-xdist==3.6.1
python-bitcoinlib==0.12.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1706886791323/work
PyYAML==6.0.2
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
rich @ file:///home/conda/feedstock_root/build_artifacts/rich_1733342254348/work/dist
ruamel.yaml @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml_1736248096639/work
ruamel.yaml.clib @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml.clib_1728724494498/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.2
soupsieve @ file:///home/conda/feedstock_root/build_artifacts/soupsieve_1693929250441/work
tomli @ file:///home/conda/feedstock_root/build_artifacts/tomli_1731426768275/work
tqdm==4.67.1
types-psutil @ file:///home/conda/feedstock_root/build_artifacts/types-psutil_1734758640583/work
types-pytz @ file:///home/conda/feedstock_root/build_artifacts/types-pytz_1727940197272/work
types-PyYAML==6.0.12.20250402
types-requests==2.32.0.20250328
types-tqdm @ file:///home/conda/feedstock_root/build_artifacts/types-tqdm_1713338972121/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1733188668063/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1727140567071/work
untokenize @ file:///home/conda/feedstock_root/build_artifacts/untokenize_1723841446890/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1734859416348/work
zstandard==0.23.0
